requirements.txt
python-dotenv
Flask
Flask-Login
Flask-SocketIO
Flask-WTF
Jinja2
MarkupSafe
SQLAlchemy
Werkzeug
WTForms
Flask-SQLAlchemy
Flask-Migrate
Flask-Mail


